+++ lib/core/runtime.ml
@@ -107,6 +107,13 @@ module Make (Runtime : Required.RUNTIME) = struct
                         (function
                           | Ok x -> continue k x | Error err -> runtimec err)
                         (Runtime.create_directory ~on:filesystem path))
+              | Eff.Yocaml_erase_file (filesystem, path) ->
+                  Some
+                    (fun (k : (a, _) continuation) ->
+                      Runtime.bind
+                        (function
+                          | Ok () -> continue k () | Error err -> runtimec err)
+                        (Runtime.erase_file ~on:filesystem path))
               | Eff.Yocaml_is_directory (filesystem, path) ->
                   Some
                     (fun (k : (a, _) continuation) ->